backend/requirements/base.txt
# Core Framework
fastapi==0.115.7
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
starlette==0.45.3

# Database
SQLAlchemy==2.0.37
alembic==1.14.1
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7

# Document Processing
python-docx==1.1.2
docxtpl==0.19.1
docxcompose==1.4.0
Pillow==11.1.0

# Utilities
python-dotenv==1.0.1
python-json-logger==2.0.4
aiofiles==24.1.0
httpx==0.28.1

# Monitoring & Logging
sentry-sdk==2.21.0 